// packages/components/src/atoms/typography/text/enum.ts
export enum TextType {
	ABBREVIATION = 'ABBREVIATION',
	BOLD = 'BOLD',
	CITATION = 'CITATION',
	DELETED = 'DELETED',
	EMPHASIS = 'EMPHASIS',
	INSERTED = 'INSERTED',
	ITALIC = 'ITALIC',
	KEYBOARD = 'KEYBOARD',
	LABEL = 'LABEL',
	MARK = 'MARK',
	PARAGRAPH = 'PARAGRAPH',
	SAMPLE = 'SAMPLE',
	STRIKETHROUGH = 'STRIKETHROUGH',
	SUBSCRIPT = 'SUBSCRIPT',
	SUPERSCRIPT = 'SUPERSCRIPT',
	UNDERLINE = 'UNDERLINE',
}